Lua: load/save architecture-independent bytecode
If you run the Intel version of Crawl and then the PowerPC version,
you will get "Lua error: global_prelude: bad header in precompiled
chunk" seven times during the start of Crawl.

This patch corrects the issue by making the bytecode architecture
independent.

+++ crawl-ref/source/util/lua/src/ldump.c
@@ -62,12 +62,12 @@ static void DumpString(const TString* s, DumpState* D)
 {
  if (s==NULL || getstr(s)==NULL)
  {
-  size_t size=0;
+  unsigned int size=0;
   DumpVar(size,D);
  }
  else
  {
-  size_t size=s->tsv.len+1;		/* include trailing '\0' */
+  unsigned int size=s->tsv.len+1;		/* include trailing '\0' */
   DumpVar(size,D);
   DumpBlock(getstr(s),size,D);
  }
